Commit 65affdda28c56762bf7ecd0bfaa2dbc18fc160ca forgot to check if the directory exists before deleting, this commit is a fix to it

+++ init.php
@@ -57,6 +57,6 @@ function removeDirectory($path) {
 		rmdir($path);
 		return;
 	}
-
-removeDirectory(HACKADEMIC_PATH."/installation");
+if(file_exists (HACKADEMIC_PATH."/installation"))
+	removeDirectory(HACKADEMIC_PATH."/installation");
 ?>
